refactor: create MenuProductModel
