+++ src/main/java/kitchenpos/menus/domain/vo/MenuProducts.java
@@ -23,4 +23,5 @@ public class MenuProducts {
     public int size() {
         return menuProducts.size();
     }
+
 }
